Add support to handle 429 responses from a rate limited gitlab instance
Gitlab may respond with a 429 error and a message of 'Retry later'.  The
response body in this case has content-type 'text/plain'. This change
adds support for 429 errors and processes the body of responses with
content formated as 'text/plain'.

Add tests for the changes to Gitlab:Error:ResponseError.build_error_message