Delegate Hash methods in ObjectifiedHash
Assuming this class is meant to mimic a normal Ruby Hash if the sent
message doesn't exist as a data key, we need to also delegate methods
that the normal Hash responds to (e.g., `include?`).

This also corrects the method signature for `method_missing` to include
the remaining arguments and optional block.