Thoroughly remove the client prefix for pagination response
Fixes #494

There can be a mismatch in the API endpoint and the pagination endpoint. In our use-case, we whitelist the Gitlab API on gitlab-internal.domain.com, whereas the web interface gitlab.domain.com cannot contact the API. However, Gitlab pagination links point to gitlab.domain.com which does not work. There seem to be some other use cases in #494 and #521.

This change is to trust that the gem is configured with the appropriate endpoint. The assumption is that only the domain prefix would be different and not the api version-that could cause additional issues and should fail.