Prevent error obfuscation on json parse errors
Sometimes the Gitlab API responds with an error code with incorrect
content-type headers. (application/json header with text/plain body)
This causes a Parsing error to be raised instead
of the api error (e.g. throttling). This change updates
ResponseError to ignore parsing errors so that the original api error
is still raised.