Respect HTTP 429 and Retry-After header
When a 429 is received, sleep for the length of time indicated by Retry-After, then try again.

Try up to 3 times before giving up, although the count can be configured with :ratelimit_retries if you want more or less

If no Retry-After is sent by the server, defaults to 2 seconds (something is better than nothing). That should only happen with quite old and unsupported versions of GitLab