Delete group hook
https://docs.gitlab.com/ee/api/groups.html#delete-group-hook

The API returns status and empty body. The library converts the empty body into an empty Gitlab::ObjectifiedHash. In test, the stub converts an empty body to False. Decided to represent all of this by making the fixture accuractly represent the response body and have the test method call out the discrepancy.