Add body_as_json configuration parameter
Allow the user to configure the client to use
a Content-Type of application/json for all requests
with a body unless the request explicitly provides
its own Content-Type header.

Using application/json provides a broader range
of compatibility with the GitLab API, especially
for calls where the body expects more complex
data structures.

This defaults to false for backwards compatibility.